backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.9.3
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
